requirements.txt
pywebview>=4.4.0
pywebview[qt]>=4.4.0
